improve submitter and watcher scripts

+++ encd-atac-pl_watcher-v3.sh
@@ -34,21 +34,20 @@ if [[ "$counter" -ne "$max_samp_count" ]]; then
             set +e # Disable the exit on error option so that the script can continue if the croo post-processing script fails
             # Run the croo post-processing script
             echo "Block 1 watcher script"
-            if . /home/suffi.azizan/scratchspace/pipeline_scripts/atac-seq-workflow-scripts/croo_processing_module.sh "${analysis_id}" "${pl_raw_output_root_dir}/${analysis_id}" "${croo_output_root_dir}"; then
+            captured_err=$(bash /home/suffi.azizan/scratchspace/pipeline_scripts/atac-seq-workflow-scripts/croo_processing_module.sh "${analysis_id}" "${pl_raw_output_root_dir}/${analysis_id}" "${croo_output_root_dir}" | tee "/home/suffi.azizan/scratchspace/pipeline_scripts/atac-seq-workflow-scripts/output_files/logs/encd-atac-pl_watcher_${analysis_id}.log" | grep -oE "RSYNC_ERROR|CROO_ERROR")
+            # check if captured_err array is empty
+            if [[ -z "$captured_err" ]]; then
                 echo "Croo post-processing script completed successfully."
                 echo "Processed files have been transferred to remote storage Odin."
                 # Remove the sample directory in the dataset-specific directory to save space
-                echo "Removing sample subdirectories in ${analysis_id} from the pipeline raw output directory..."
+                echo "WARNING: Removing sample subdirectories in ${analysis_id} from the pipeline raw output directory..."
                 find "${pl_raw_output_root_dir}/${analysis_id}" -maxdepth 1 -mindepth 1 -type d -exec rm -rf {} \;
                 echo "All sample subdirectories have been copied to Odin and removed from Gekko."
                 echo "Submitting the next batch of jobs..."
                 source /home/suffi.azizan/scratchspace/pipeline_scripts/atac-seq-workflow-scripts/encd-atac-pl_submitter-v3.sh "${analysis_id}" "${dataset_json_dir}" "${pl_raw_output_root_dir}" "${croo_output_root_dir}" "${counter}"
-
             else
-                echo "Croo post-processing script failed for this batch of jobs."
+                echo "Croo post-processing script failed for this batch of jobs. Please check the log file for more details."
             fi
-        
-    
         elif (( finish_counts != MAX_JOBS )) && [[ $(qstat -u suffi.azizan | grep -c "CAPER_${analysis_id:0:3}") -eq 0 ]]; then
             echo "Something is wrong with the current pipeline jobs..."
             set +e # Disable the exit on error option so that the script can continue if the croo post-processing script fails
@@ -58,25 +57,24 @@ if [[ "$counter" -ne "$max_samp_count" ]]; then
                 # Run the croo post-processing script
                 echo "Block 2 watcher script"
                 echo "Proceeding with croo post-processing script..."
-                if . /home/suffi.azizan/scratchspace/pipeline_scripts/atac-seq-workflow-scripts/croo_processing_module.sh "${analysis_id}" "${pl_raw_output_root_dir}/${analysis_id}" "${croo_output_root_dir}"; then
+                captured_err=$(bash /home/suffi.azizan/scratchspace/pipeline_scripts/atac-seq-workflow-scripts/croo_processing_module.sh "${analysis_id}" "${pl_raw_output_root_dir}/${analysis_id}" "${croo_output_root_dir}" | tee "/home/suffi.azizan/scratchspace/pipeline_scripts/atac-seq-workflow-scripts/output_files/logs/encd-atac-pl_watcher_${analysis_id}.log" | grep -oE "RSYNC_ERROR|CROO_ERROR")
+                # check if captured_err array is empty
+                if [[ -z "$captured_err" ]]; then
                     echo "Croo post-processing script completed successfully."
                     echo "Files successfully processed have been transferred to remote storage Odin."
                     # Remove the sample directory in the dataset-specific directory to save space
                     echo "Removing sample subdirectories in ${analysis_id} from the pipeline raw output directory..."
                     find "${pl_raw_output_root_dir}/${analysis_id}" -maxdepth 1 -mindepth 1 -type d -exec rm -rf {} \;
                     echo "All processed subdirectories have been copied to Odin and removed from Gekko."
-
                     echo "Submitting the next batch of jobs..."
                     source /home/suffi.azizan/scratchspace/pipeline_scripts/atac-seq-workflow-scripts/encd-atac-pl_submitter-v3.sh "${analysis_id}" "${dataset_json_dir}" "${pl_raw_output_root_dir}" "${croo_output_root_dir}" "${counter}"
-
                 else
-                    echo "Croo post-processing script failed for this batch of jobs."
+                    echo "Croo post-processing script failed for this batch of jobs. Please check the log file for more details."
                 fi
             else
                 echo "No successful jobs found."
             fi
-            
-
+            echo "No jobs are running now. Exiting watcher script..."
         elif (( finish_counts != MAX_JOBS )) && [[ $(qstat -u suffi.azizan | grep -c "CAPER_${analysis_id:0:3}") -ne 0 ]]; then
             echo "The current batch of submitted jobs are still running. Will check again in 1 hour."
             echo "Number of metadata.json files indicating finished jobs: $finish_counts"
@@ -103,7 +101,9 @@ else
             set +e # Disable the exit on error option so that the script can continue if the croo post-processing script fails
             # Run the croo post-processing script
             echo "Block 3 watcher script"
-            if . /home/suffi.azizan/scratchspace/pipeline_scripts/atac-seq-workflow-scripts/croo_processing_module.sh "${analysis_id}" "${pl_raw_output_root_dir}/${analysis_id}" "${croo_output_root_dir}"; then
+            captured_err=$(bash /home/suffi.azizan/scratchspace/pipeline_scripts/atac-seq-workflow-scripts/croo_processing_module.sh "${analysis_id}" "${pl_raw_output_root_dir}/${analysis_id}" "${croo_output_root_dir}" | tee "/home/suffi.azizan/scratchspace/pipeline_scripts/atac-seq-workflow-scripts/output_files/logs/encd-atac-pl_watcher_${analysis_id}.log" | grep -oE "RSYNC_ERROR|CROO_ERROR")
+            # check if captured_err array is empty
+            if [[ -z "$captured_err" ]]; then
                 echo "Croo post-processing script completed successfully."
                 echo "Processed files have been transferred to remote storage Odin."
                 # Remove the sample directory in the dataset-specific directory to save space
@@ -111,7 +111,7 @@ else
                 find "${pl_raw_output_root_dir}/${analysis_id}" -maxdepth 1 -mindepth 1 -type d -exec rm -rf {} \;
                 echo "All sample subdirectories have been copied to Odin and removed from Gekko."
             else
-                echo "Croo post-processing script failed for this batch of jobs."
+                echo "Croo post-processing script failed for this batch of jobs. Please check the log file for more details."
             fi
             echo "All job runs have finished. Exiting watcher script..."
             echo "Workflow is done."
@@ -123,7 +123,9 @@ else
                 # Run the croo post-processing script
                 echo "Block 4 watcher script"
                 echo "Proceeding with croo post-processing script..."
-                if . /home/suffi.azizan/scratchspace/pipeline_scripts/atac-seq-workflow-scripts/croo_processing_module.sh "${analysis_id}" "${pl_raw_output_root_dir}/${analysis_id}" "${croo_output_root_dir}"; then
+                captured_err=$(bash /home/suffi.azizan/scratchspace/pipeline_scripts/atac-seq-workflow-scripts/croo_processing_module.sh "${analysis_id}" "${pl_raw_output_root_dir}/${analysis_id}" "${croo_output_root_dir}" | tee "/home/suffi.azizan/scratchspace/pipeline_scripts/atac-seq-workflow-scripts/output_files/logs/encd-atac-pl_watcher_${analysis_id}.log" | grep -oE "RSYNC_ERROR|CROO_ERROR")
+                # check if captured_err array is empty
+                if [[ -z "$captured_err" ]]; then
                     echo "Croo post-processing script completed successfully."
                     echo "Files successfully processed have been transferred to remote storage Odin."
                     # Remove the sample directory in the dataset-specific directory to save space
@@ -131,12 +133,12 @@ else
                     find "${pl_raw_output_root_dir}/${analysis_id}" -maxdepth 1 -mindepth 1 -type d -exec rm -rf {} \;
                     echo "All processed subdirectories have been copied to Odin and removed from Gekko."
                 else
-                    echo "Croo post-processing script failed for this batch of jobs."
+                    echo "Croo post-processing script failed for this batch of jobs. Please check the log file for more details."
                 fi
             else
                 echo "No successful jobs found."
             fi
-            echo "All job runs have finished. Exiting watcher script..."
+            echo "No jobs are running now. Exiting watcher script..."
         elif (( finish_counts != remainder )) && [[ $(qstat -u suffi.azizan | grep -c "CAPER_${analysis_id:0:3}") -ne 0 ]]; then
             echo "The current batch of submitted jobs are still running. Will check again in 1 hour."
             echo "Number of metadata.json files indicating finished jobs: $finish_counts"
